+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ## Updates of Existing Functions
 
-* Updated `assertions_filter_cond` by adding a `gsub()` argument   on line 520 so that error messages return `filter_add` instead of  `enexpr(filter_add)`.
+* Updated arg_name parameter in `assert_filter_cond`, `assert_symbol`, and `assert_expr` with  `gsub("^enexpr\\((\\w+)\\)$", "\\1", rlang::caller_arg(arg))` argument so that  error messages return a quoted argument without `enexp()`, e.g. `filter_add` instead of `enexpr(filter_add)`.
 
 * Error messaging throughout the package have been updated from `rlang::abort()` to `cli::cli_abort()`. As a part of the update, each of the `assert_*()` functions have new arguments `assert_*(message, arg_name, call, class).` (#367)